Basically raw commit after tinyist changes (to package names)
A chunk of code I've accumulated over the last few years.  I am trying to keep
only the most basic stuff that I would recreate basically by typing almost the
exact same thing and not anything specific to a particular job or that might
be considered valuable IP or in any sense fancy: no AWS stuff, for example.

I already know this stuff doesn't work: I have fixes that are in the next
commit: this is the stuff in the index and that stuff is ready to be added to
it, i.e. the difference between 'git diff --cached' and 'git diff'.