Whoops -- update build.gradle with necessary config for running JUnit 5 tests.
I was annoyed that Gradle output no longer seems to say how many tests
completed successfully, so I went and looked at the test build report.  It
said 0!  So a quick Googling got this config change.  Unfortunately, it
*still* doesn't say anything about # of successful tests on the commandline,
but the build report says 26 successful, so I guess roll with that for now.